+++ db/comdb2.h
@@ -3591,5 +3591,6 @@ extern int gbl_bpfunc_auth_gen;
 
 void dump_client_sql_data(struct reqlogger *logger, int do_snapshot);
 
+int backout_schema_changes(struct ireq *iq, tran_type *tran);
 int bplog_schemachange(struct ireq *iq, blocksql_tran_t *tran, void *err);
 #endif /* !INCLUDED_COMDB2_H */
